Mocking-CustomError

+++ src/controllers/products.controller.js
@@ -6,10 +6,22 @@ import {
   deleteProductService,
   getProductsByCategoryService,
   orderProductByPriceService,
+  mockingProductsService,
 } from "../service/products.service.js";
+
+import { CustomError } from "../service/customError.service.js";
+import { userAuthError } from "../service/authError.service.js";
+import { EError } from "../enums/EError.js";
 
 export const getProductsController = async (req, res) => {
   try {
+    if (!req.user) {
+      throw CustomError.createError({
+        name: "User Not Logged",
+        message: "El usuario debe estar logueado para acceder a esta ruta",
+        errorCode: EError.UserNotLogged,
+      });
+    }
     const products = await getProductService();
     const { limit } = req.query;
     if (limit) {
@@ -34,6 +46,13 @@ export const getProductByIdController = async (req, res) => {
 
 export const addProductController = async (req, res) => {
   try {
+    if (req.user.role === "user") {
+      throw CustomError.createError({
+        name: "is not admin",
+        message: userAuthError(req.user),
+        errorCode: EError.userAuthError,
+      });
+    }
     const { code, title, description, price, thumbnail, stock, category } =
       req.body;
     const product = await addProductService({
@@ -53,6 +72,13 @@ export const addProductController = async (req, res) => {
 
 export const updateProductController = async (req, res) => {
   try {
+    if (req.user.role === "user") {
+      throw CustomError.createError({
+        name: "is not admin",
+        message: userAuthError(req.user),
+        errorCode: EError.userAuthError,
+      });
+    }
     const { id } = req.params;
     const { code, title, description, price, thumbnail, stock, category } =
       req.body;
@@ -75,6 +101,13 @@ export const updateProductController = async (req, res) => {
 
 export const deleteProductController = async (req, res) => {
   try {
+    if (req.user.role === "user") {
+      throw CustomError.createError({
+        name: "is not admin",
+        message: userAuthError(req.user),
+        errorCode: EError.userAuthError,
+      });
+    }
     const { id } = req.params;
     const product = await deleteProductService(id);
     res.send({ status: "success", payload: product });
@@ -110,6 +143,16 @@ export const orderProductByPriceController = async (req, res) => {
   } catch (err) {
     console.log(err);
 
+    res.status(404).send({ status: "error", error: `${err}` });
+  }
+};
+
+export const mockingProductsController = async (req, res) => {
+  try {
+    const mocking = await mockingProductsService();
+    res.send({ status: "success", payload: mocking });
+  } catch (err) {
+    console.log(err);
     res.status(404).send({ status: "error", error: `${err}` });
   }
 };
